Added README specfying steps necessary for build, clang 15 compatibility

+++ src/speed.c
@@ -13,6 +13,6 @@ void Speed_onNormal()
     if (!Config_showSpeed())
         return;
 
-    sprintf(speedLine, "S %.1f", (f32) gMarioStates->forwardVel);
+    sprintf(speedLine, "S %.1f", gMarioStates->forwardVel);
     TextManager_addLine(speedLine, 1);
 }
